A problem faced using replace
Here are the two folders that make the different outputs. “compare_and_check_all_files.py” shows the comparison, but I added the “1 - Diff Checker.pdf” made by the diffchecker website. It is more than anything odd for me. But I will see.
I am also sure it is due to the “replace” part. I check if there is any alternative to replace to do it. The last time I fixed it, the fixture was based on reordering, but I will search for a better solution this time.