Testing a dictionary as state value.
I cannot find the error. Since the state value is not limited to any data type,  I am not sure what is the source of the error. The state value is a dictionary and it works for one time-step only.  In the second time step, the first key is considered as a state instead of the dictionary.